add WorkspacePdfFileView
